refactor: make examples for usable for each other
