+++ examples/server-client/server.go
@@ -1,0 +1,18 @@
+package main
+
+import (
+	"log"
+
+	"github.com/ravsii/elgo"
+	"github.com/ravsii/elgo/examples/player"
+	"github.com/ravsii/elgo/socket"
+)
+
+func main() {
+	pool := elgo.NewPool()
+	pool.AddPlayer(
+		player.New("Example 1", 0),
+		player.New("Example 2", 0))
+
+	log.Fatal(socket.Listen(25561, pool))
+}
